requirements.txt
# Enterprise Knowledge Engine - Production RAG System
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
pika>=1.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.2.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
langchain-experimental>=0.0.50

# Vector DB
chromadb>=0.4.22

# Document loaders
pypdf>=4.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
pytesseract>=0.3.10
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
markdown>=3.5.0

# Utilities
fpdf2>=2.7.0
